* boom analyzer was halving the amplitudes for the FFT since it was artificially twice as high. Fixed * --warning in blockanalyzer * make turbine use standard transform again
svn path=/trunk/kdeextragear-1/amarok/; revision=336449

+++ src/analyzers/blockanalyzer.cpp
@@ -55,7 +55,7 @@ BlockAnalyzer::resizeEvent( QResizeEvent *e )
     m_columns = limit( uint(double(width()+1) / (WIDTH+1)), MAX_COLUMNS, MIN_COLUMNS );
     m_scope.resize( m_columns );
 
-    if( (uint)e->oldSize().height() != height() ) //this block speeds up window resizes vastly
+    if( e->oldSize().height() != height() ) //this block speeds up window resizes vastly
     {
         //NOTE height should only be set once! but it tends to get set many times when Qt is setting up
         //     the layout of the toolBar, what a waste of cycles!
